Backend/app/utils/auth.py
import time
from datetime import datetime, timedelta, timezone
from typing import Optional
from uuid import UUID

import jwt
from fastapi import Depends, HTTPException, Request
from fastapi.security import HTTPBearer
from firebase_admin import auth as firebase_auth
from sqlalchemy.orm import Session

from app.core.config import settings
from app.core.constant.messages import MessageConstants
from app.db import get_db
from app.models.user import User


def verify_firebase_token(id_token: str) -> dict:
    """
    Verify Firebase ID token and return decoded token payload.
    """
    try:
        # Ensure id_token is a string and not bytes
        if isinstance(id_token, bytes):
            id_token = id_token.decode("utf-8")

        # Basic validation - JWT tokens should have 3 parts separated by dots
        if not id_token or not isinstance(id_token, str):
            raise ValueError("Invalid token format: token must be a non-empty string")

        token_parts = id_token.split(".")
        if len(token_parts) != 3:
            raise ValueError(f"Invalid JWT format: expected 3 parts, got {len(token_parts)}")

        # Verify the token with Firebase
        time.sleep(2)  # Small delay to avoid rapid-fire requests in case of retries
        decoded_token = firebase_auth.verify_id_token(id_token)
        return decoded_token
    except ValueError as e:
        raise HTTPException(
            status_code=401,
            detail=f"{MessageConstants.INVALID_GOOGLE_TOKEN_FORMAT}: {str(e)}"
        )
    except Exception as e:
        raise HTTPException(
            status_code=401,
            detail=f"{MessageConstants.INVALID_GOOGLE_TOKEN}: {str(e)}"
        )


def get_firebase_user_info(id_token: str) -> dict:
    """
    Get user information from Firebase ID token.
    """
    decoded_token = verify_firebase_token(id_token)
    return {
        "uid": decoded_token.get("uid"),
        "email": decoded_token.get("email"),
        "name": decoded_token.get("name"),
        "picture": decoded_token.get("picture"),
    }


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + (expires_delta or timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire, "type": "access"})
    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm="HS256")


def create_refresh_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + (expires_delta or timedelta(minutes=settings.REFRESH_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire, "type": "refresh"})
    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm="HS256")


def verify_token(token: str):
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
        return payload
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None
    except Exception:
        return None


def get_current_user_from_token(token: str):
    payload = verify_token(token)

    if not payload:
        return None

    token_type = payload.get("type")

    if token_type != "access":
        return None

    user_id = payload.get("sub")
    return user_id


class JWTBearer(HTTPBearer):
    def __init__(self, auto_error: bool = True):
        super().__init__(auto_error=auto_error, scheme_name="BearerAuth")

    async def __call__(self, request: Request) -> str:
        authorization = request.headers.get("Authorization")
        if not authorization:
            raise HTTPException(status_code=403, detail=MessageConstants.AUTH_REQUIRED)

        # Parse the authorization header
        try:
            scheme, credentials = authorization.split(" ", 1)
        except ValueError:
            raise HTTPException(
                status_code=403,
                detail=MessageConstants.INVALID_AUTH_SCHEME
            )

        if scheme.lower() != "bearer":
            raise HTTPException(
                status_code=403,
                detail=MessageConstants.INVALID_AUTH_SCHEME
            )

        # Return the token - verification will happen in get_current_user
        return credentials


jwt_bearer = JWTBearer()


def get_current_user(token: str = Depends(jwt_bearer), db: Session = Depends(get_db)):
    """
    Extract user information from JWT token.
    """
    try:
        # Strip Bearer prefix if present (for direct calls to this function)
        import re
        token = re.sub(r'^[Bb]earer\s+', '', token).strip()

        user_id = get_current_user_from_token(token)
        if not user_id:
            raise HTTPException(status_code=401, detail=MessageConstants.INVALID_TOKEN)

        user = db.query(User).filter(User.id == UUID(user_id)).first()
        if not user:
            raise HTTPException(status_code=404, detail=MessageConstants.USER_NOT_FOUND)

        return user
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(
            status_code=401,
            detail=MessageConstants.TOKEN_VERIFICATION_FAILED
        ) from e
