Backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.13
pydantic[email]>=2.0.0
alembic>=1.11.0
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.3.0
pytest-asyncio>=1.2.0
typing-extensions>=4.5.0
passlib[bcrypt]>=1.7.4
faker>=20.0.0
PyJWT>=2.10.1
firebase-admin>=7.1.0
redis>=4.5.0
aioredis>=2.0.0
celery>=5.3.0

tenacity>=9.1.2


requests>=2.31.0

minio>=7.1.0
python-docx>=1.1.0
python-multipart>=0.0.6
markdown2>=2.4.0
colorama>=0.4.6